Drop db_mysql80_platform
