+++ .travis.yml
@@ -37,7 +37,7 @@ matrix:
         - sudo mysql_upgrade
         - mysql --version
       env:
-        - PLATFORM=mysql:8.0
+        - PLATFORM=mysql
         - SERVER_NAME=127.0.0.1
         - DB_NAME=workspace
         - DB_USER=root
